pyproject.toml
[project]
name = "my-crawl4ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "atproto>=0.0.63",
    "crawl4ai>=0.7.7",
    "openrouter>=0.0.19",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
    "yt-dlp[curl-cffi,default]>=2025.11.12",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]

[tool.pytest.ini_options]
pythonpath = "src"
markers = [
    "integration: mark a test as an integration test",
]